The commit message type is a "fix" as it specifies that unnecessary debug statements have been removed.(fix(commit-msg): Remove unnecessary debug statements
This commit type is a fix(commit-msg) which describes the change made in the commit as removing unnecessary debug statements from the code. The commit message indicates that the commit is fixing an issue related to the commit message itself, specifically removing the debug statements for improved clarity. The commit message includes the commit ID or reference number (4096) for further tracking or reference.): chore(commit-msg): remove unnecessary debug statements (4096) - Semantic Version: patch